move pages to src/pages
