+++ src/pages/api/product/[category].ts
