+++ src/pages/fruits-distractions.tsx
